+++ Services/JobsService.cs
@@ -19,24 +19,32 @@ namespace contractorserver.Services
         return _repo.GetAll();
     }
 
-    internal Job GetById(int id)
+    internal Job GetById(int id, Profile userInfo)
     {
       var data =  _repo.GetById(id);
-      if(data == null)
+      if(data.CreatorId != userInfo.Id || data == null)
       {
-        throw new Exception("Invalid ID");
-      } 
+        throw new Exception("No Permission or Invalid ID!");
+      }
       return data;
     }
 
-    internal Job Create(Job newJob)
+    internal Job Create(Job newJob, Profile userInfo)
     {
+      if(newJob.CreatorId != userInfo.Id)
+      {
+        throw new Exception("Invalid Permissions");
+      }
       return _repo.Create(newJob);
     }
 
-    internal Job Edit(Job updated)
+    internal Job Edit(Job updated, Profile userInfo)
     {
-      var data = GetById(updated.Id);
+      if(updated.CreatorId != userInfo.Id)
+      {
+        throw new Exception("Invalid Permissions to Edit");
+      }
+      var data = GetById(updated.Id, userInfo);
       updated.Location = updated.Location != null ? updated.Location : data.Location;
       updated.Description = updated.Description != null ? updated.Description : data.Description;
       updated.ContactPhone = updated.ContactPhone > 0 ? updated.ContactPhone : data.ContactPhone;
@@ -45,9 +53,14 @@ namespace contractorserver.Services
       return _repo.Edit(updated);
     }
 
-    internal String Delete(int id)
+    internal String Delete(int id, Profile userInfo)
     {
-      var data = GetById(id);
+
+      var data = GetById(id, userInfo);
+      if(data.CreatorId != userInfo.Id)
+      {
+        throw new Exception("Invalid Permissions to Delete");
+      }
       _repo.Delete(id);
       return "Successfully Deleted!";
     }
